Fixed possible deadlock when game-mode flag is being used
When joining to region with game-mode flag set while initializing flag handlers PlayerGameModeChangeEvent is fired causing the plugin to try check the current status of fly flag to decide whenever can they fly or not. This causes issues as the intializers are being fired from the guava's LoadingCache causing the get to block, tho Guava actually prevents deadlocks and throws exception.

The workaround is to check if the session is present at the moment and if not then scheduale it for next tick, otherwise we can continue